requirements.txt
# Backend
fastapi
uvicorn
python-multipart
requests

# Frontend
streamlit

# ML / DL
transformers==4.45.0
huggingface-hub==0.35.3
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
numpy==1.26.4
pandas==2.3.3
joblib==1.5.2
tqdm==4.67.1

# Image Processing
pillow==11.3.0

# Visualization (optional but commonly used)
matplotlib==3.10.6
seaborn==0.13.2

# Utilities
PyYAML==6.0.3
regex==2025.9.18
filelock==3.19.1
networkx==3.5
sympy==1.14.0
typing_extensions
packaging
python-dateutil
pytz
tzdata

# HTTP
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# CORS / ASGI
starlette
